Initial managed multi-GPU support (#1830)
Related issue: #1583

* update test due to bug in conda
The version of conda previously used in this unit test (in the base-images-testing repo) had a bug with conda-forge, so solving the env started to suddenly fail. The image used has been updated in that repo and then the version bumped here since a newer version of python was installed in the base.